+++ packages/bundle/src/adaptiveCards/Attachment/AdaptiveCardBuilder.ts
@@ -119,11 +119,11 @@ export default class AdaptiveCardBuilder {
   }
 
   addCommonHeaders(content: ICommonContent) {
-    const { richCardWrapTitle } = this.styleOptions;
+    const { richCardTitleOmitHeadingRole, richCardWrapTitle } = this.styleOptions;
     this.addTextBlock(content.title, {
+      ...(richCardTitleOmitHeadingRole ? {} : { style: 'heading' }),
       color: TextColor.Default,
       size: TextSize.Medium,
-      style: 'heading',
       weight: TextWeight.Bolder,
       wrap: richCardWrapTitle
     });