+++ CHANGELOG.md
@@ -19,6 +19,11 @@ Legends:
 
 ## [Unreleased]
 
+### Added
+
+- Added `styleOptions.richCardTitleOmitHeadingRole` (default `false`) to opt out of `style: 'heading'` on rich card titles, in PR [#5839](https://github.com/microsoft/BotFramework-WebChat/pull/5839), by [@cjennison](https://github.com/cjennison)
+
+
 ## [4.19.1] - 2026-06-09
 
 ### Changed